Move author copyright to individual files
Instead of trying to handle the list of authors separately to the file
copyright, remove AUTHORS.md and change the copyright in the files to my
name instead (as I'm the only author thus far).  Any future contributors
can add their own copyright notices directly to the files they
contribute to.

Also take the opportunity to reduce the size of the headers by removing
the GPL boilerplate from each file and replacing it with a
SPDX-License-Identifier header line.